requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pyyaml>=6.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.1
